requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
anthropic>=0.25.0
PyYAML>=6.0
rumps>=0.4.0
PyObjC>=10.0
